Refactor: Split manage.test.ts into modular per-operation files (warren-e73a)
