+++ .mulch/expertise/projects.jsonl
@@ -15,3 +15,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-18T15:39:16.293Z","evidence":{"seeds":"warren-b960","commit":"a6d06704d905896d80e949fda3b8822d9ebddb57"},"content":"src/projects/refresh.test.ts (warren-b960, pl-d4d6 step 2) tests the defaultPreservePlot wrapper end-to-end against a real on-disk clone: each test mkdtemp's a clone root, writes .plot/ fixtures, and mocks SpawnFn so the 'reset --hard origin/<ref>' handler rm -rf's the .plot/ dir to simulate git wiping the working tree. The snapshot-then-restore wrapper must put bytes back. Six tests: events.jsonl byte-equality across refresh, local-'done' status.json winning over origin-'active', no-snapshot-dir for projects without .plot/, 100-consecutive-refresh leak smoke (counts warren-plot-snapshot-* under os.tmpdir() before/after), .index.db* skip (mx-239786 rebuild), recursive nested-file preservation.","id":"mx-bf632c"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-18T16:18:18.705Z","evidence":{"seeds":"warren-ee9d","commit":"5b07e3659d3df8bc6c05d5a4f3d23871f0d514cb"},"content":"Host-side state that must survive refreshProjectClone resets (uncommitted working-tree files written by warren itself, e.g. .plot/ appender writes) must go through an injectable preserve-wrapper that snapshots to mkdtemp before 'git reset --hard' and restores after. Pattern: defaultPreservePlot in src/projects/refresh.ts — gate on a pre-reset existence probe (do NOT trust post-reset features detection, the reset itself can flip hasPlot), snapshot regular files recursively skipping derived/gitignored state (e.g. .index.db*), restore with snapshot-wins-on-conflict, clean up tmp dir in finally. Projects without the feature directory must short-circuit before any fs work so the no-feature path stays byte-identical to the pre-wrapper refresh.","id":"mx-973fc7"}
 {"type":"pattern","classification":"tactical","recorded_at":"2026-05-23T06:39:56.034Z","evidence":{"seeds":"warren-af9e","commit":"a4f77bf8285c4bf6eec9270110b0c23f66123bdd"},"name":"plot-snapshot-merge-on-refresh","description":"defaultPreservePlot merges rather than blindly overwriting: snapshotPlotDir captures only flat plot-*.events.jsonl and plot-*.json files (SPEC §11.O flat layout); mergePlotSnapshot uses dedup-append for events and field-level merge (remote base + host-side status overlay) for JSON. Remote attachments/intent/events survive the refresh while host-appender status transitions are preserved.","files":[".claude/settings.local.json",".seeds/issues.jsonl","src/projects/refresh.test.ts","src/projects/refresh.ts"],"id":"mx-4f5c9a"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T04:12:49.091Z","evidence":{"commit":"f4c1b37dfd84f260b9bae97c40c51159ae715a43"},"content":"Split large manage.test.ts file (763 lines) into modular, per-operation test files under the 500-line budget ceiling, alongside a shared manage.test-helpers.ts helper file for common mocks, recording caches, and configurations.","id":"mx-0dfb89"}
