+++ scripts/file-size-budgets.json
@@ -17,7 +17,6 @@
 		"src/diagnostics/checks.ts": 616,
 		"src/plan-runs/coordinator.ts": 587,
 		"src/plots/aggregate.ts": 506,
-		"src/projects/manage.test.ts": 763,
 		"src/projects/refresh.test.ts": 709,
 		"src/runs/pause.test.ts": 616,
 		"src/runs/pr.test.ts": 604,
